+++ sdk/agrifood/arm-agrifood/src/agriFoodMgmtClient.ts
@@ -42,7 +42,7 @@ export class AgriFoodMgmtClient extends coreClient.ServiceClient {
   /**
    * Initializes a new instance of the AgriFoodMgmtClient class.
    * @param credentials Subscription credentials which uniquely identify client subscription.
-   * @param subscriptionId The ID of the target subscription.
+   * @param subscriptionId The ID of the target subscription. The value must be an UUID.
    * @param options The parameter options
    */
   constructor(
@@ -66,7 +66,7 @@ export class AgriFoodMgmtClient extends coreClient.ServiceClient {
       credential: credentials
     };
 
-    const packageDetails = `azsdk-js-arm-agrifood/1.0.0-beta.4`;
+    const packageDetails = `azsdk-js-arm-agrifood/1.0.0-beta.5`;
     const userAgentPrefix =
       options.userAgentOptions && options.userAgentOptions.userAgentPrefix
         ? `${options.userAgentOptions.userAgentPrefix} ${packageDetails}`