+++ sdk/agrifood/arm-agrifood/src/operations/privateEndpointConnections.ts
@@ -284,7 +284,7 @@ const createOrUpdateOperationSpec: coreClient.OperationSpec = {
     Parameters.farmBeatsResourceName,
     Parameters.privateEndpointConnectionName
   ],
-  headerParameters: [Parameters.accept, Parameters.contentType],
+  headerParameters: [Parameters.contentType, Parameters.accept],
   mediaType: "json",
   serializer
 };